requirements.txt
stramlit
pandas
streamlit-option-menu
google.generativeai